+++ packages/client/src/Call.ts
@@ -36,8 +36,8 @@ import type {
   DeleteCallRequest,
   DeleteCallResponse,
   EndCallResponse,
+  GetCallReportResponse,
   GetCallResponse,
-  GetCallStatsResponse,
   GetOrCreateCallRequest,
   GetOrCreateCallResponse,
   GoLiveRequest,
@@ -151,6 +151,7 @@ import {
   PromiseWithResolvers,
   promiseWithResolvers,
 } from './helpers/promise';
+import { GetCallStatsResponse } from './gen/shims';
 
 /**
  * An object representation of a `Call`.
@@ -237,6 +238,7 @@ export class Call {
   public readonly streamClient: StreamClient;
   private sfuClient?: StreamSfuClient;
   private sfuClientTag = 0;
+  private unifiedSessionId?: string;
 
   private readonly reconnectConcurrencyTag = Symbol('reconnectConcurrencyTag');
   private reconnectAttempts = 0;
@@ -373,7 +375,7 @@ export class Call {
         const currentUserId = this.currentUserId;
         if (currentUserId && blockedUserIds.includes(currentUserId)) {
           this.logger('info', 'Leaving call because of being blocked');
-          await this.leave({ reason: 'user blocked' }).catch((err) => {
+          await this.leave({ message: 'user blocked' }).catch((err) => {
             this.logger('error', 'Error leaving call after being blocked', err);
           });
         }
@@ -554,10 +556,7 @@ export class Call {
   /**
    * Leave the call and stop the media streams that were published by the call.
    */
-  leave = async ({
-    reject,
-    reason = 'user is leaving the call',
-  }: CallLeaveOptions = {}) => {
+  leave = async ({ reject, reason, message }: CallLeaveOptions = {}) => {
     await withoutConcurrency(this.joinLeaveConcurrencyTag, async () => {
       const callingState = this.state.callingState;
       if (callingState === CallingState.LEFT) {
@@ -577,7 +576,7 @@ export class Call {
 
       if (callingState === CallingState.RINGING && reject !== false) {
         if (reject) {
-          await this.reject('decline');
+          await this.reject(reason ?? 'decline');
         } else {
           // if reject was undefined, we still have to cancel the call automatically
           // when I am the creator and everyone else left the call
@@ -600,7 +599,9 @@ export class Call {
       this.publisher?.dispose();
       this.publisher = undefined;
 
-      await this.sfuClient?.leaveAndClose(reason);
+      await this.sfuClient?.leaveAndClose(
+        message ?? reason ?? 'user is leaving the call',
+      );
       this.sfuClient = undefined;
       this.dynascaleManager.setSfuClient(undefined);
 
@@ -612,6 +613,7 @@ export class Call {
       this.leaveCallHooks.forEach((hook) => hook());
       this.initialized = false;
       this.hasJoinedOnce = false;
+      this.unifiedSessionId = undefined;
       this.ringingSubject.next(false);
       this.cancelAutoDrop();
       this.clientStore.unregisterCall(this);
@@ -1180,7 +1182,6 @@ export class Call {
       state: this.state,
       connectionConfig,
       logTag: String(this.sfuClientTag),
-      clientDetails,
       enableTracing,
       onUnrecoverableError: (reason) => {
         this.reconnect(WebsocketReconnectStrategy.REJOIN, reason).catch(
@@ -1209,7 +1210,6 @@ export class Call {
         connectionConfig,
         publishOptions,
         logTag: String(this.sfuClientTag),
-        clientDetails,
         enableTracing,
         onUnrecoverableError: (reason) => {
           this.reconnect(WebsocketReconnectStrategy.REJOIN, reason).catch(
@@ -1236,6 +1236,7 @@ export class Call {
 
     this.sfuStatsReporter?.stop();
     if (statsOptions?.reporting_interval_ms > 0) {
+      this.unifiedSessionId ??= sfuClient.sessionId;
       this.sfuStatsReporter = new SfuStatsReporter(sfuClient, {
         clientDetails,
         options: statsOptions,
@@ -1244,6 +1245,7 @@ export class Call {
         microphone: this.microphone,
         camera: this.camera,
         state: this.state,
+        unifiedSessionId: this.unifiedSessionId,
       });
       this.sfuStatsReporter.start();
     }
@@ -1533,7 +1535,7 @@ export class Call {
       const { reconnectStrategy: strategy, error } = e;
       if (strategy === WebsocketReconnectStrategy.UNSPECIFIED) return;
       if (strategy === WebsocketReconnectStrategy.DISCONNECT) {
-        this.leave({ reason: 'SFU instructed to disconnect' }).catch((err) => {
+        this.leave({ message: 'SFU instructed to disconnect' }).catch((err) => {
           this.logger('warn', `Can't leave call after disconnect request`, err);
         });
       } else {
@@ -1702,6 +1704,12 @@ export class Call {
         await this.publisher.publish(audioTrack, TrackType.SCREEN_SHARE_AUDIO);
         trackTypes.push(TrackType.SCREEN_SHARE_AUDIO);
       }
+    }
+
+    if (track.kind === 'video') {
+      // schedules calibration report - the SFU will use the performance stats
+      // to adjust the quality thresholds as early as possible
+      this.sfuStatsReporter?.scheduleOne(3000);
     }
 
     await this.updateLocalStreamState(mediaStream, ...trackTypes);
@@ -2338,12 +2346,15 @@ export class Call {
 
     // 0 means no auto-drop
     if (timeoutInMs <= 0) return;
-
     this.dropTimeout = setTimeout(() => {
       // the call might have stopped ringing by this point,
       // e.g. it was already accepted and joined
       if (this.state.callingState !== CallingState.RINGING) return;
-      this.leave({ reject: true, reason: 'timeout' }).catch((err) => {
+      this.leave({
+        reject: true,
+        reason: 'timeout',
+        message: `ringing timeout - ${this.isCreatedByMe ? 'no one accepted' : `user didn't interact with incoming call screen`}`,
+      }).catch((err) => {
         this.logger('error', 'Failed to drop call', err);
       });
     }, timeoutInMs);
@@ -2394,10 +2405,26 @@ export class Call {
    *
    * @param callSessionID the call session ID to retrieve statistics for.
    * @returns The call stats.
+   * @deprecated use `call.getCallReport` instead.
+   * @internal
    */
   getCallStats = async (callSessionID: string) => {
     const endpoint = `${this.streamClientBasePath}/stats/${callSessionID}`;
     return this.streamClient.get<GetCallStatsResponse>(endpoint);
+  };
+
+  /**
+   * Retrieve call report. If the `callSessionID` is not specified, then the
+   * report for the latest call session is retrieved. If it is specified, then
+   * the report for that particular session is retrieved if it exists.
+   *
+   * @param callSessionID the optional call session ID to retrieve statistics for
+   * @returns the call report
+   */
+  getCallReport = async (callSessionID: string = '') => {
+    const endpoint = `${this.streamClientBasePath}/report`;
+    const params = callSessionID !== '' ? { session_id: callSessionID } : {};
+    return this.streamClient.get<GetCallReportResponse>(endpoint, params);
   };
 
   /**