+++ packages/client/src/gen/video/sfu/signal_rpc/signal.ts
@@ -8,6 +8,7 @@ import {
   ICETrickle,
   InputDevices,
   PeerType,
+  PerformanceStats,
   RTMPIngress,
   TrackInfo,
   TrackType,
@@ -158,13 +159,39 @@ export interface SendStatsRequest {
    */
   rtmp?: RTMPIngress;
   /**
-   * @generated from protobuf field: string subscriber_rtc_stats = 13;
+   * @deprecated
+   * @generated from protobuf field: string subscriber_rtc_stats = 13 [deprecated = true];
    */
   subscriberRtcStats: string;
   /**
-   * @generated from protobuf field: string publisher_rtc_stats = 14;
+   * @deprecated
+   * @generated from protobuf field: string publisher_rtc_stats = 14 [deprecated = true];
    */
   publisherRtcStats: string;
+  /**
+   * @generated from protobuf field: string rtc_stats = 15;
+   */
+  rtcStats: string;
+  /**
+   * Encode stats for the publisher
+   *
+   * @generated from protobuf field: repeated stream.video.sfu.models.PerformanceStats encode_stats = 16;
+   */
+  encodeStats: PerformanceStats[];
+  /**
+   * Decode stats for the subscriber
+   *
+   * @generated from protobuf field: repeated stream.video.sfu.models.PerformanceStats decode_stats = 17;
+   */
+  decodeStats: PerformanceStats[];
+  /**
+   * user_session id can change during reconnects, this helps us to
+   * identify the user across reconnects and should remain consistent until the user explicitly
+   * disconnects, is kicked or the call is ended.
+   *
+   * @generated from protobuf field: string unified_session_id = 18;
+   */
+  unifiedSessionId: string;
 }
 /**
  * @generated from protobuf message stream.video.sfu.signal.SendStatsResponse
@@ -531,6 +558,27 @@ class SendStatsRequest$Type extends MessageType<SendStatsRequest> {
       {
         no: 14,
         name: 'publisher_rtc_stats',
+        kind: 'scalar',
+        T: 9 /*ScalarType.STRING*/,
+      },
+      { no: 15, name: 'rtc_stats', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
+      {
+        no: 16,
+        name: 'encode_stats',
+        kind: 'message',
+        repeat: 1 /*RepeatType.PACKED*/,
+        T: () => PerformanceStats,
+      },
+      {
+        no: 17,
+        name: 'decode_stats',
+        kind: 'message',
+        repeat: 1 /*RepeatType.PACKED*/,
+        T: () => PerformanceStats,
+      },
+      {
+        no: 18,
+        name: 'unified_session_id',
         kind: 'scalar',
         T: 9 /*ScalarType.STRING*/,
       },